Created TradeGood classes and implemented Market generation
Classes for all TradeGoods have been added.
The Market now generates 2 HashMaps: one maps a good to its availability
on the planet (null means planet cannot sell it), the other maps a good
to its price on the planet.
Good prices are randomly generated according to the formula noted in the
generatePrices() method.

**** TODO: add feature to display prices to console for testing/demo
purposes