Implemented save and load (almost)
Saving and loading works using a hybrid of text file/serializable
methods. Current SolarSystem and Player are saved and loaded from a text
file, the Player's Ship is saved and loaded from a bin file.

The only thing missing is to save/load the Universe which I should
hopefully have finished tonight.